+++ CHANGELOG.md
@@ -7,4 +7,5 @@
  -
 
 ## Fixes
+ - Add `-preset` option
  - Fix rendering audio when speed is less than 0.5x